ready

+++ JBtest/JBapi/views.py
@@ -32,7 +32,8 @@ def post_list(request):
 		posts.delete()
 		images_dir = os.getcwd() + '\\JBapi\\image_files'
 		for f in os.listdir(images_dir):
-			os.remove(os.path.join(images_dir, f))
+			if f != 'images.txt':
+				os.remove(os.path.join(images_dir, f))
 		return Response(status=status.HTTP_204_NO_CONTENT)
 
 
@@ -75,6 +76,7 @@ def post_detail(request, pk):
 		if elements_num == 1:
 			image_file_name = url_of_element.split('/')[-1]
 			image_path = os.getcwd() + '\\JBapi\\image_files\\' + image_file_name
+			print(image_path)
 			if os.path.isfile(image_path):
 				os.remove(image_path)
 		post.delete()
